refactor run; move to stage.utils
