+++ dvc/repo/experiments/run.py
@@ -2,7 +2,7 @@ import logging
 from typing import Iterable, Optional
 
 from dvc.repo import locked
-from dvc.utils.cli_parse import loads_params_from_cli
+from dvc.utils.cli_parse import loads_params
 
 logger = logging.getLogger(__name__)
 
@@ -28,7 +28,7 @@ def run(
         return repo.experiments.reproduce_queued(jobs=jobs)
 
     if params:
-        params = loads_params_from_cli(params)
+        params = loads_params(params)
     return repo.experiments.reproduce_one(
         targets=targets, params=params, tmp_dir=tmp_dir, **kwargs
     )
