+++ dvc/utils/cli_parse.py
@@ -2,7 +2,7 @@ from collections import defaultdict
 from typing import Any, Dict, Iterable, List
 
 
-def parse_params_from_cli(path_params: Iterable[str]) -> Dict[str, List[str]]:
+def parse_params(path_params: Iterable[str]) -> Dict[str, List[str]]:
     """Normalizes the shape of params from the CLI to dict."""
     from dvc.dependency.param import ParamsDependency
 
@@ -17,9 +17,7 @@ def parse_params_from_cli(path_params: Iterable[str]) -> Dict[str, List[str]]:
     return ret
 
 
-def loads_params_from_cli(
-    path_params: Iterable[str],
-) -> Dict[str, Dict[str, Any]]:
+def loads_params(path_params: Iterable[str],) -> Dict[str, Dict[str, Any]]:
 
     """Loads the content of params from the cli as Python object."""
     from ruamel.yaml import YAMLError
@@ -28,7 +26,7 @@ def loads_params_from_cli(
 
     from .serialize import loads_yaml
 
-    normalized_params = parse_params_from_cli(path_params)
+    normalized_params = parse_params(path_params)
     ret: Dict[str, Dict[str, Any]] = defaultdict(dict)
 
     for path, param_keys in normalized_params.items():
